+++ .idea/.name
@@ -1,1 +1,1 @@
-SBS
+sbs
