--- .idea/libraries/native_libs.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="native-libs">
-    <CLASSES>
-      <root url="jar://$PROJECT_DIR$/sbs/build/native-libs/native-libs.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
